chore: minor fixes to auth provider component
