+++ src/providers/auth/index.tsx
@@ -1,4 +1,5 @@
 import React, { FC, PropsWithChildren, memo, useState, useEffect } from 'react';
+import { compose } from 'redux';
 
 import service from '../../services/auth';
 
@@ -6,13 +7,16 @@ import Context from './context';
 
 const AuthProvider: FC<PropsWithChildren<any>> = ({ children }) => {
   const [isInitialised, setIsInitialised] = useState(false);
+  const [isAuthenticated, setIsAuthenticated] = useState(false);
 
   const init = async () => {
     try {
-      await service.init(true);
+      const authenticated = await service.init();
+
       setIsInitialised(true);
-    } catch (error) {
-      console.error(error);
+      setIsAuthenticated(authenticated);
+    } catch (e) {
+      // TODO: handle service errors and log them to Sentry
     }
   };
 
@@ -20,12 +24,13 @@ const AuthProvider: FC<PropsWithChildren<any>> = ({ children }) => {
     init();
   }, []);
 
-  return isInitialised ? (
+  return isInitialised && isAuthenticated ? (
     <Context.Provider value={{ service }}>{children}</Context.Provider>
   ) : (
     <></>
   );
 };
 
-export default memo(AuthProvider);
-export { withAuth, ServiceProps as Props } from './hook';
+export default compose<FC>(memo)(AuthProvider);
+export { withAuth } from './hoc';
+export type { ServiceProps as Props } from './hoc';
